Remove testing button from headers, and move it to dev space
